+++ src/memory.c
@@ -1,16 +1,17 @@
 #include "memory.h"
 #include "logger.h"
 
+#include <stdlib.h>
 #include <string.h>
 
 struct MemoryStats {
-    size_t allocated;
-    size_t allocated_tagged[MEMORY_TAG_MAX_TAGS];
+    u64 allocated;
+    u64 allocated_tagged[MEMORY_TAG_MAX_TAGS];
 };
 
 static struct MemoryStats stats;
 
-void *s_alloc(size_t size, memory_tag tag)
+void *s_alloc(u64 size, memory_tag tag)
 {
     if (tag == MEMORY_TAG_UNKNOWN)
     {
@@ -22,13 +23,12 @@ void *s_alloc(size_t size, memory_tag tag)
 
     void *mem = malloc(size);
 
-
     s_zeroMemory(mem, size);
 
     return mem;
 }
 
-void s_free(void *mem, size_t size, memory_tag tag)
+void s_free(void *mem, u64 size, memory_tag tag)
 {
     if (tag == MEMORY_TAG_UNKNOWN)
     {
@@ -41,7 +41,7 @@ void s_free(void *mem, size_t size, memory_tag tag)
     free(mem);
 }
 
-void s_zeroMemory(void *mem, size_t size)
+void s_zeroMemory(void *mem, u64 size)
 {
     memset(mem, 0, size);
 }
